pptx-to-vector/requirements.txt
psycopg2==2.9.10
sentence_transformers==3.4.1
python-dotenv==1.0.0